# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  rating: number;
  stock: number;
  description: string;
  image: string;
}

export interface TimelineStep {
  label: string;
  date: string;
  completed: boolean;
  current: boolean;
}

export interface Order {
  id: string;
  product: string;
  productImage: string;
  date: string;
  status: string;
  expectedDelivery: string;
  amount: number;
  customer: string;
  trackingNumber: string;
  timeline: TimelineStep[];
}

export interface Return {
  id: string;
  orderId: string;
  product: string;
  reason: string;
  status: string;
  pickupDate: string;
  refundStatus: string;
  amount: number;
}

export interface ReturnInput {
  /** @minLength 1 */
  orderId: string;
  /** @minLength 1 */
  reason: string;
}

export interface CustomerProfile {
  id: string;
  name: string;
  initials: string;
  email: string;
  preferredCategory: string;
  lastPurchase: string;
  supportPreference: string;
  pastOrders: number;
  interactions: number;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export interface CategoryMetric {
  label: string;
  value: number;
  color: string;
}

export interface ToolMetric {
  name: string;
  calls: number;
  share: number;
}

export interface Activity {
  id: string;
  tool: string;
  title: string;
  detail: string;
  time: string;
  tone: string;
}

export interface Dashboard {
  totalConversations: number;
  resolvedByAi: number;
  avgResponseTime: number;
  activeCustomers: number;
  conversationTrend: TrendPoint[];
  issueCategories: CategoryMetric[];
  resolutionTrend: TrendPoint[];
  toolUsage: ToolMetric[];
  activity: Activity[];
}

export interface SupportMessageInput {
  /** @minLength 1 */
  message: string;
}

export type SupportMessageResponseResult = { [key: string]: unknown };

export interface SupportMessageResponse {
  id: string;
  reply: string;
  category: string;
  tool: string;
  toolLabel: string;
  resultSummary: string;
  result: SupportMessageResponseResult;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProductsParams = {
q?: string;
category?: string;
maxPrice?: number;
minRating?: number;
};