# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: shopassist
    description: ShopAssist AI demo operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /shopassist/products:
    get:
      operationId: listProducts
      tags: [shopassist]
      summary: List products
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /shopassist/orders:
    get:
      operationId: listOrders
      tags: [shopassist]
      summary: List customer orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /shopassist/orders/{orderId}:
    get:
      operationId: getOrder
      tags: [shopassist]
      summary: Get a single order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /shopassist/returns:
    get:
      operationId: listReturns
      tags: [shopassist]
      summary: List returns
      responses:
        "200":
          description: Returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Return"
    post:
      operationId: createReturn
      tags: [shopassist]
      summary: Create a return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnInput"
      responses:
        "201":
          description: Return created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Return"
        "400":
          description: Invalid return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shopassist/customer-profile:
    get:
      operationId: getCustomerProfile
      tags: [shopassist]
      summary: Get the demo customer profile
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
  /shopassist/dashboard:
    get:
      operationId: getDashboard
      tags: [shopassist]
      summary: Get agent dashboard metrics
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /shopassist/support/messages:
    post:
      operationId: sendSupportMessage
      tags: [shopassist]
      summary: Send a message to the AI support agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportMessageInput"
      responses:
        "200":
          description: Agent response with friendly tool activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessageResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, category, price, rating, stock, description, image]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        rating: { type: number }
        stock: { type: integer }
        description: { type: string }
        image: { type: string }
    Order:
      type: object
      required: [id, product, productImage, date, status, expectedDelivery, amount, customer, trackingNumber, timeline]
      properties:
        id: { type: string }
        product: { type: string }
        productImage: { type: string }
        date: { type: string }
        status: { type: string }
        expectedDelivery: { type: string }
        amount: { type: number }
        customer: { type: string }
        trackingNumber: { type: string }
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineStep"
    TimelineStep:
      type: object
      required: [label, date, completed, current]
      properties:
        label: { type: string }
        date: { type: string }
        completed: { type: boolean }
        current: { type: boolean }
    Return:
      type: object
      required: [id, orderId, product, reason, status, pickupDate, refundStatus, amount]
      properties:
        id: { type: string }
        orderId: { type: string }
        product: { type: string }
        reason: { type: string }
        status: { type: string }
        pickupDate: { type: string }
        refundStatus: { type: string }
        amount: { type: number }
    ReturnInput:
      type: object
      required: [orderId, reason]
      properties:
        orderId: { type: string, minLength: 1 }
        reason: { type: string, minLength: 1 }
    CustomerProfile:
      type: object
      required: [id, name, initials, email, preferredCategory, lastPurchase, supportPreference, pastOrders, interactions]
      properties:
        id: { type: string }
        name: { type: string }
        initials: { type: string }
        email: { type: string }
        preferredCategory: { type: string }
        lastPurchase: { type: string }
        supportPreference: { type: string }
        pastOrders: { type: integer }
        interactions: { type: integer }
    Dashboard:
      type: object
      required: [totalConversations, resolvedByAi, avgResponseTime, activeCustomers, conversationTrend, issueCategories, resolutionTrend, toolUsage, activity]
      properties:
        totalConversations: { type: integer }
        resolvedByAi: { type: number }
        avgResponseTime: { type: number }
        activeCustomers: { type: integer }
        conversationTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        issueCategories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryMetric"
        resolutionTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        toolUsage:
          type: array
          items:
            $ref: "#/components/schemas/ToolMetric"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    TrendPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
    CategoryMetric:
      type: object
      required: [label, value, color]
      properties:
        label: { type: string }
        value: { type: number }
        color: { type: string }
    ToolMetric:
      type: object
      required: [name, calls, share]
      properties:
        name: { type: string }
        calls: { type: integer }
        share: { type: number }
    Activity:
      type: object
      required: [id, tool, title, detail, time, tone]
      properties:
        id: { type: string }
        tool: { type: string }
        title: { type: string }
        detail: { type: string }
        time: { type: string }
        tone: { type: string }
    SupportMessageInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
    SupportMessageResponse:
      type: object
      required: [id, reply, category, tool, toolLabel, resultSummary, result, createdAt]
      properties:
        id: { type: string }
        reply: { type: string }
        category: { type: string }
        tool: { type: string }
        toolLabel: { type: string }
        resultSummary: { type: string }
        result:
          type: object
          additionalProperties: true
        createdAt: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }