# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: string;
  tool: string;
  title: string;
  detail: string;
  time: string;
  tone: string;
}