# lib/api-zod/src/generated/types/customerProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerProfile {
  id: string;
  name: string;
  initials: string;
  email: string;
  preferredCategory: string;
  lastPurchase: string;
  supportPreference: string;
  pastOrders: number;
  interactions: number;
}