# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { CategoryMetric } from './categoryMetric';
import type { ToolMetric } from './toolMetric';
import type { TrendPoint } from './trendPoint';

export interface Dashboard {
  totalConversations: number;
  resolvedByAi: number;
  avgResponseTime: number;
  activeCustomers: number;
  conversationTrend: TrendPoint[];
  issueCategories: CategoryMetric[];
  resolutionTrend: TrendPoint[];
  toolUsage: ToolMetric[];
  activity: Activity[];
}