# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineStep } from './timelineStep';

export interface Order {
  id: string;
  product: string;
  productImage: string;
  date: string;
  status: string;
  expectedDelivery: string;
  amount: number;
  customer: string;
  trackingNumber: string;
  timeline: TimelineStep[];
}