# lib/api-zod/src/generated/types/return.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Return {
  id: string;
  orderId: string;
  product: string;
  reason: string;
  status: string;
  pickupDate: string;
  refundStatus: string;
  amount: number;
}