# lib/api-zod/src/generated/types/returnInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReturnInput {
  /** @minLength 1 */
  orderId: string;
  /** @minLength 1 */
  reason: string;
}