# lib/api-zod/src/generated/types/supportMessageResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SupportMessageResponseResult } from './supportMessageResponseResult';

export interface SupportMessageResponse {
  id: string;
  reply: string;
  category: string;
  tool: string;
  toolLabel: string;
  resultSummary: string;
  result: SupportMessageResponseResult;
  createdAt: string;
}