# lib/api-zod/src/generated/types/timelineStep.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineStep {
  label: string;
  date: string;
  completed: boolean;
  current: boolean;
}