# lib/api-zod/src/generated/types/toolMetric.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ToolMetric {
  name: string;
  calls: number;
  share: number;
}